Break out Firsts into its own module.
In retrospect, this should probably have been done earlier.
