+++ src/lib/mod.rs
@@ -38,6 +38,7 @@ use std::fmt;
 
 
 mod ast;
+mod firsts;
 pub mod grammar;
 mod yacc_parser;
 mod stategraph;
